+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 
 IMPROVEMENTS:
 
-- Improve ali_vpc and make it idempotence ([#134](https://github.com/alibaba/ansible-provider/pull/134))
+- Improve ali_vswitch and make it idempotence ([#136](https://github.com/alibaba/ansible-provider/pull/136))
+- Improve ali_vpc and make it idempotence ([#135](https://github.com/alibaba/ansible-provider/pull/135))
 
 ## 1.8.0 (12 November, 2018)